Fix the audio ES input found in audit run 11
- MPEG audio and AC3 frame_time is samples / sample rate, not byte length
  / bit rate. At 44.1 kHz header 0 - the cut plan's grid - read 26.083
  instead of 26.122 ms and every kept segment lost ~20 ms (10 segments
  over 1 h: 198 ms, the drift column showing 10). 48 kHz is unchanged
  (H1). MPEG-2/2.5 Layer I/II frames were parsed at half their length, so
  16-24 kHz MP2 was refused (H5).
- A header that yields no frame is skipped and the MPEG header list goes
  on; it used to end there (H3).
- The open dialog and the automatic audio search offer exactly
  TTAVTypes::readableAudioSuffixes (mpa, mp2, mp3, ac3); AAC, E-AC3 and
  DTS were offered and then refused. TODO records the missing parsers;
  CLAUDE.md no longer lists .aac as input (H2, user decision).
- Silence and audio-change markers map their audio time through
  videoFrameForTime with the video's extra frames, like the anomaly scan:
  they were one frame early per extra frame, audio-change markers of a
  44.1 kHz AC3 track 48.8 s early at 10:00 (H4, user decision).
  abs_frame_time is a double; its float sum was 80 ms off after 10 min.

Gate: new audio_es_input (15 checks, material built with ffmpeg); 7 of
the first 13 fail on the old code, the H4 checks fail with the old
formulas in place.

+++ avstream/ttavheader.h
@@ -98,7 +98,10 @@ public:
   //protected:
   long   position;  // header offset ???
   float  frame_time;
-  float  abs_frame_time;
+  // Start time of the frame in ms, the running sum of the frame_times
+  // before it. double: a float sum of 34.83 ms (AC3 at 44.1 kHz) is 80 ms
+  // off after 10 minutes (audit run 11).
+  double abs_frame_time;
   int    frame_length;
   int    bit_rate;
   int    sample_rate;
